Don't use WAL mode & extend schema with metadata
The WAL mode is persistent and only works with sqlite >= 3.7, which isn't
great for our distribution use case. It does help performance a bit with
synchronous=NORMAL by only requiring an fsync on checkpoint (every 1000 pages
or so), but we might as well set synchronous=OFF for dumps.

This patch also adds revision metadata in the schema, but does not actually
fill this yet. We'll need to fetch this information separately, or extract
(some of) it from the HTML head. The former will work for other formats like
wikitext or JSON as well, so should be preferable.